Add files via upload
Made it glide a little bit but it's incredibly basic so I don't think it's right. There are a few equations I tried using that didn't seem to work but probably will once we get the camera up and running. Also I used a coroutine to attempt to make the glide work and it wouldn't stop right so I had to use StopAllCoroutines(). There is also a TestShooter function for testing the grapple target but I commented it out cause I forgot how to do it like in 180 and I can just move the player towards the target to make it work.