Homing Enemy Finish, More work on shooting enemy, etc
Finished work on the homing enemy. When it hits the player, it seems to hit a lot but I also programmed invincibility for the player so that it won't immediately delete all its health. The shooting enemy is currently not finished, so don't use it yet. I'm having trouble getting it to deal damage. I started a bit of work on the goal watch tower. It's very plain. For now it just displays a panel that says you made it to the goal.